Create a common mainline robot base point for easy use as a starting point for all future robots. (#1)
Create a common mainline robot base point for easy use as a starting point for all future robots. This code is largely independent of specific robots and can be used as the starting commit for code development each season. Some assumptions are made such as the presence of a camera and generic code for various components that may or may not be present exists. The main branch will continue to be updated only with generic functionality while specific branches will contain specific code for each season.